Snake-Game
It is a simple snake game where a snake moves around a box and trying eat food when the snake eat succesfully the length and movement of the snake increses and when the snake touch to his own body then the game will be over. It is made by using simple HTML , CSS and JavaScript.